Split classes into other files
